Hook up new lexer to parser
