+++ library/Amy/Syntax.hs
@@ -2,6 +2,7 @@ module Amy.Syntax
   ( module X
   ) where
 
+import Amy.Syntax.Lexer as X (lexer)
 import Amy.Syntax.Monad as X
 import Amy.Syntax.Parser as X
 import Amy.Syntax.Pretty as X
