+++ library/Amy/Syntax/Monad.hs
@@ -8,19 +8,25 @@ module Amy.Syntax.Monad
   , currentIndentation
   , assertIndented
   , assertSameIndentation
+  , indentedBlock
+  , indentedBlockNonEmpty
   ) where
 
 import Control.Applicative (Alternative)
+import qualified Control.Applicative.Combinators.NonEmpty as CNE
 import Control.Monad.State.Strict
+import Data.List.NonEmpty (NonEmpty)
 import Data.Monoid ((<>))
 import Data.Text (Text, pack, unpack)
 import Data.Void (Void)
 import Text.Megaparsec hiding (State)
 
-newtype AmyParser a = AmyParser (StateT Int (Parsec Void Text) a)
-  deriving (Functor, Applicative, Alternative, MonadPlus, Monad, MonadState Int, MonadParsec Void Text)
+import Amy.Syntax.Lexer
 
-runAmyParser :: AmyParser a -> Parsec Void Text a
+newtype AmyParser a = AmyParser (StateT Int (Parsec Void AmyTokens) a)
+  deriving (Functor, Applicative, Alternative, MonadPlus, Monad, MonadState Int, MonadParsec Void AmyTokens)
+
+runAmyParser :: AmyParser a -> Parsec Void AmyTokens a
 runAmyParser (AmyParser action) = evalStateT action 0
 
 withBlockIndentation :: AmyParser a -> AmyParser a
@@ -55,3 +61,19 @@ checkIndentation msg rel = do
   col <- unPos . sourceColumn <$> getPosition
   current <- currentIndentation
   guard (col `rel` current) <?> unpack (msg <> " " <> pack (show current))
+
+-- A block is defined as a group of things that are at the same indentation
+-- level. Block items can also be separated by semicolons.
+indentedBlock
+  :: AmyParser a
+  -> AmyParser [a]
+indentedBlock p =
+  fmap concat $ withBlockIndentation $ many $
+    assertSameIndentation *> p `sepBy1` semiColon
+
+indentedBlockNonEmpty
+  :: AmyParser a
+  -> AmyParser (NonEmpty a)
+indentedBlockNonEmpty p =
+  fmap join $ withBlockIndentation $ CNE.some $
+    assertSameIndentation *> p `CNE.sepBy1` semiColon
